Moving predictions to own file
